+++ libs/zhi-blog-api/src/lib/IWebApi.ts
@@ -62,6 +62,13 @@ interface IWebApi extends IBlogApi {
   getMetaData(): Promise<any>
 
   /**
+   * 退出 Web 认证会话
+   *
+   * @returns Promise<boolean> 退出结果
+   */
+  logoutWebAuth(): Promise<boolean>
+
+  /**
    * 内容预处理：预处理平台无法兼容的文本内容
    *
    * @param post 文章对象