+++ libs/zhi-siyuan-api/CHANGELOG.md
@@ -1,5 +1,13 @@
 # zhi-siyuan-api
 
+## 2.35.4
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - zhi-blog-api@1.78.0
+
 ## 2.35.3
 
 ### Patch Changes